refactor: remove unused state props and reorg style module
