+++ packages/browser-ui/src/inspector/views/viewMap.tsx
@@ -2,6 +2,7 @@
 import Errors from "../../inspector/views/Errors";
 import CompGraph from "./CompGraph";
 import Frames from "./Frames";
+import Opt from "./Opt";
 import Settings from "./Settings";
 import ShapeView from "./ShapeView";
 
@@ -10,7 +11,7 @@ const viewMap = {
   errors: Errors,
   shapes: ShapeView,
   // mod: Mod, // NOTE: mod tab temporarily deprecated
-  //"optimization status": Opt,
+  "optimization status": Opt,
   compGraph: CompGraph,
   settings: Settings,
 };
