+++ packages/core/src/types/style.ts
@@ -317,6 +317,7 @@ export type GPIDecl<T> = ASTNode<T> & {
   properties: PropertyDecl<T>[];
 };
 
+// TODO: limit `below` and `above` to `FieldPath`
 export type ILayering<T> = ASTNode<T> & {
   tag: "Layering";
   below: Path<T>;
